requirements.txt
absl-py==2.2.2
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
deepface==0.0.93
Django==5.2.1
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
keras==3.9.2
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mtcnn==1.0.0
namex==0.0.9
npm==0.1.1
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optional-django==0.1.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
